+++ consensus/types/src/signed_execution_payload_envelope.rs
@@ -54,37 +54,6 @@ impl<E: EthSpec> SignedExecutionPayloadEnvelope<E> {
             Self::NextFork(ref signed) => ExecutionPayloadEnvelopeRef::NextFork(&signed.message),
         }
     }
-
-    /// Verify `self.signature`.
-    ///
-    /// The `parent_state` is the post-state of the beacon block with
-    /// block_root = self.message.beacon_block_root
-    pub fn verify_signature(
-        &self,
-        parent_state: &BeaconState<E>,
-        genesis_validators_root: Hash256,
-        spec: &ChainSpec,
-    ) -> Result<bool, BeaconStateError> {
-        let domain = spec.get_domain(
-            parent_state.current_epoch(),
-            Domain::BeaconBuilder,
-            &parent_state.fork(),
-            genesis_validators_root,
-        );
-        let pubkey = parent_state
-            .validators()
-            .get(self.message().builder_index() as usize)
-            .and_then(|v| {
-                let pk: Option<PublicKey> = v.pubkey.decompress().ok();
-                pk
-            })
-            .ok_or_else(|| {
-                BeaconStateError::UnknownValidator(self.message().builder_index() as usize)
-            })?;
-        let message = self.message().signing_root(domain);
-
-        Ok(self.signature().verify(&pubkey, message))
-    }
 }
 
 impl<'de, E: EthSpec> ContextDeserialize<'de, ForkName> for SignedExecutionPayloadEnvelope<E> {
